refactor(utils): move mana_icon_factory + mana_resources to services/mana_icon_service (#443)
ManaIconFactory and ManaIconResources depend on wx, so they're not generic
utils. Move both as one unit into services/mana_icon_service/, split into
factory, bitmap_renderer (mixin), cache, and resources. ManaIconFactory
now composes BitmapRendererMixin to preserve the single-class call-site
surface.

Updates all 18 widget callers plus tests/test_mana_icon_factory.py.
Keeps the deferred TYPE_CHECKING imports at widgets/panels/mana_rich_text_ctrl
deferred. utils/mana_icon_factory.py and utils/mana_resources.py deleted.
