+++ services/mana_icon_service/resources.py
