+++ widgets/frames/app_frame/protocol.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Any, Protocol
 import wx
 import wx.lib.agw.flatnotebook as fnb
 
-from utils.mana_icon_factory import ManaIconFactory
+from services.mana_icon_service import ManaIconFactory
 from widgets.buttons.toolbar_buttons import ToolbarButtons
 from widgets.panels.card_inspector_panel import CardInspectorPanel
 from widgets.panels.card_panel import CardPanel
