+++ widgets/panels/card_box_panel/properties.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, Any
 
 import wx
 
-from utils.mana_icon_factory import ManaIconFactory
+from services.mana_icon_service import ManaIconFactory
 
 if TYPE_CHECKING:
     from widgets.panels.card_box_panel.protocol import CardBoxPanelProto
