+++ widgets/panels/card_inspector_panel/frame.py
@@ -28,7 +28,7 @@ from utils.constants import (
     PADDING_XL,
     SUBDUED_TEXT,
 )
-from utils.mana_icon_factory import ManaIconFactory
+from services.mana_icon_service import ManaIconFactory
 from widgets.panels.card_image_display import CardImageDisplay
 from widgets.panels.card_inspector_panel.handlers import CardInspectorPanelHandlersMixin
 from widgets.panels.card_inspector_panel.properties import CardInspectorPanelPropertiesMixin
