+++ widgets/panels/card_inspector_panel/protocol.py
@@ -9,7 +9,7 @@ import wx
 
 from services.image_service import CardImageRequest
 from utils.card_data import CardDataManager
-from utils.mana_icon_factory import ManaIconFactory
+from services.mana_icon_service import ManaIconFactory
 from widgets.panels.card_image_display import CardImageDisplay
 from widgets.panels.mana_rich_text_ctrl import ManaSymbolRichCtrl
