+++ widgets/panels/mana_rich_text_ctrl/frame.py
@@ -47,7 +47,7 @@ from widgets.panels.mana_rich_text_ctrl.properties import (
 )
 
 if TYPE_CHECKING:
-    from utils.mana_icon_factory import ManaIconFactory
+    from services.mana_icon_service import ManaIconFactory
 
 
 # Three-tone frame matching the native Win11 dark-mode TextCtrl outline,
